+++ mod-audit-server/src/test/java/org/folio/builder/service/RequestRecordBuilderTest.java
@@ -17,6 +17,7 @@ import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.rest.jaxrs.model.LogRecord;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
@@ -25,8 +26,7 @@ import io.vertx.core.json.JsonObject;
 
 public class RequestRecordBuilderTest extends BuilderTestBase {
 
-  private static final Logger logger = LogManager.getLogger()
-    ;
+  private static final Logger logger = LogManager.getLogger();
 
   private static final String EXPECTED_CREATE_DESCRIPTION = "Type: Recall.";
   private static final String EXPECTED_EDITED_DESCRIPTION = "Type: Recall. New expiration date: 2020-10-23 00:00:00 (from: 2020-10-19 00:00:00). New fulfilment preference: Hold Shelf (from: Hold).";
@@ -82,7 +82,8 @@ public class RequestRecordBuilderTest extends BuilderTestBase {
     assertThat(requestLogRecord.getAction(), equalTo(value.getAction()));
 
     assertThat(requestLogRecord.getItems().get(0).getItemBarcode(), notNullValue());
-    assertThat(requestLogRecord.getItems().get(0).getItemBarcode(), notNullValue());
+    assertThat(requestLogRecord.getItems().get(0).getHoldingId(), notNullValue());
+    assertThat(requestLogRecord.getItems().get(0).getInstanceId(), notNullValue());
 
     assertThat(requestLogRecord.getUserBarcode(), notNullValue());
     assertThat(requestLogRecord.getLinkToIds().getUserId(), notNullValue());